+++ downscale/dataset.py
@@ -76,11 +76,12 @@ class Dataset( object ):
 		self.interp = interp
 		self.ncpus = ncpus
 		self.method = method
-		self._rotated = False
-		self._lonpc = None
-		if interp:
-			print( 'running interpolation across NAs' )
-			_ = self.interp_na( )
+
+
+		# # this may have to move to downscale
+		# if interp:
+		# 	print( 'running interpolation across NAs' )
+		# 	_ = self.interp_na( )
 	
 	# @staticmethod
 	# def transform_from_latlon( lat, lon ):
@@ -120,81 +121,81 @@ class Dataset( object ):
 			# flip each slice of the array and make a new one
 			flipped = np.array( [ np.flipud( arr ) for arr in self.ds[ self.variable ].data ] )
 			self.ds[ self.variable ] = (('time', 'lat', 'lon' ), flipped )
-	@staticmethod
-	def rotate( dat, lons, to_pacific=False ):
-		'''rotate longitudes in WGS84 Global Extent'''
-		if to_pacific == True:
-			# to 0 - 360
-			dat, lons = utils.shiftgrid( 0., dat, lons )
-		elif to_pacific == False:
-			# to -180.0 - 180.0 
-			dat, lons = utils.shiftgrid( 180., dat, lons, start=False )
-		else:
-			raise AttributeError( 'to_pacific must be boolean True:False' )
-		return dat, lons
-	@staticmethod
-	def wrap( d ):
-		return utils.xyz_to_grid( **d )
-	def interp_na( self ):
-		'''
-		np.float32
-		method = [str] one of 'cubic', 'near', 'linear'
+	# @staticmethod
+	# def rotate( dat, lons, to_pacific=False ):
+	# 	'''rotate longitudes in WGS84 Global Extent'''
+	# 	if to_pacific == True:
+	# 		# to 0 - 360
+	# 		dat, lons = utils.shiftgrid( 0., dat, lons )
+	# 	elif to_pacific == False:
+	# 		# to -180.0 - 180.0 
+	# 		dat, lons = utils.shiftgrid( 180., dat, lons, start=False )
+	# 	else:
+	# 		raise AttributeError( 'to_pacific must be boolean True:False' )
+	# 	return dat, lons
+	# @staticmethod
+	# def wrap( d ):
+	# 	return utils.xyz_to_grid( **d )
+	# def interp_na( self ):
+	# 	'''
+	# 	np.float32
+	# 	method = [str] one of 'cubic', 'near', 'linear'
 
-		return a list of dicts to pass to the xyz_to_grid in parallel
-		'''
-		from copy import copy
-		import pandas as pd
-		import numpy as np
-		from pathos.mp_map import mp_map
+	# 	return a list of dicts to pass to the xyz_to_grid in parallel
+	# 	'''
+	# 	from copy import copy
+	# 	import pandas as pd
+	# 	import numpy as np
+	# 	from pathos.mp_map import mp_map
 
-		# remove the darn scientific notation
-		np.set_printoptions( suppress=True )
-		output_dtype = np.float32
+	# 	# remove the darn scientific notation
+	# 	np.set_printoptions( suppress=True )
+	# 	output_dtype = np.float32
 		
-		# if 0-360 leave it alone
-		if ( self.ds.lon > 200.0 ).any() == True:
-			dat, lons = self.ds[ self.variable ].data, self.ds.lon
-			self._lonpc = lons
-		else:
-			# greenwich-centered rotate to 0-360 for interpolation across pacific
-			dat, lons = self.rotate( self.ds[ self.variable ].values, self.ds.lon, to_pacific=True )
-			self._rotated = True # update the rotated attribute
-			self._lonpc = lons
+	# 	# if 0-360 leave it alone
+	# 	if ( self.ds.lon > 200.0 ).any() == True:
+	# 		dat, lons = self.ds[ self.variable ].data, self.ds.lon
+	# 		self._lonpc = lons
+	# 	else:
+	# 		# greenwich-centered rotate to 0-360 for interpolation across pacific
+	# 		dat, lons = self.rotate( self.ds[ self.variable ].values, self.ds.lon, to_pacific=True )
+	# 		self._rotated = True # update the rotated attribute
+	# 		self._lonpc = lons
 
-		# mesh the lons and lats and unravel them to 1-D
-		xi, yi = np.meshgrid( self._lonpc, self.ds.lat.data )
-		lo, la = [ i.ravel() for i in (xi,yi) ]
+	# 	# mesh the lons and lats and unravel them to 1-D
+	# 	xi, yi = np.meshgrid( self._lonpc, self.ds.lat.data )
+	# 	lo, la = [ i.ravel() for i in (xi,yi) ]
 
-		# setup args for multiprocessing
-		df_list = [ pd.DataFrame({ 'x':lo, 'y':la, 'z':d.ravel() }).dropna( axis=0, how='any' ) for d in dat ]
+	# 	# setup args for multiprocessing
+	# 	df_list = [ pd.DataFrame({ 'x':lo, 'y':la, 'z':d.ravel() }).dropna( axis=0, how='any' ) for d in dat ]
 
-		args = [ {'x':np.array(df['x']), 'y':np.array(df['y']), 'z':np.array(df['z']), \
-				'grid':(xi,yi), 'method':self.method, 'output_dtype':output_dtype } for df in df_list ]
+	# 	args = [ {'x':np.array(df['x']), 'y':np.array(df['y']), 'z':np.array(df['z']), \
+	# 			'grid':(xi,yi), 'method':self.method, 'output_dtype':output_dtype } for df in df_list ]
 		
-		# # # # USE MLAB's griddata which we _can_ parallelize
-		def wrap( d ):
-			''' simple wrapper around utils.xyz_to_grid for mp_map'''
-			x = np.array( d['x'] )
-			y = np.array( d['y'] )
-			z = np.array( d['z'] )
-			xi, yi = d['grid']
-			return utils.xyz_to_grid( x, y, z, (xi,yi), interp='linear' )
-		# # # # 
+	# 	# # # # USE MLAB's griddata which we _can_ parallelize
+	# 	def wrap( d ):
+	# 		''' simple wrapper around utils.xyz_to_grid for mp_map '''
+	# 		x = np.array( d['x'] )
+	# 		y = np.array( d['y'] )
+	# 		z = np.array( d['z'] )
+	# 		xi, yi = d['grid']
+	# 		return utils.xyz_to_grid( x, y, z, (xi,yi), interp='linear' )
+	# 	# # # # 
 
-		try:
-			print( 'processing interpolation to convex hull in parallel using {} cpus.'.format( self.ncpus ) )
-			dat_list = mp_map( wrap, args, nproc=self.ncpus )
-			dat_list = [ i.data for i in dat_list ] # drop the output mask
-			dat = np.array( dat_list )
-		except:
-			print( 'processing cru re-gridding in serial due to multiprocessing issues...' )
-			dat = np.array([ wrap( **i ) for i in args ])
+	# 	try:
+	# 		print( 'processing interpolation to convex hull in parallel using {} cpus.'.format( self.ncpus ) )
+	# 		dat_list = mp_map( wrap, args, nproc=self.ncpus )
+	# 		dat_list = [ i.data for i in dat_list ] # drop the output mask
+	# 		dat = np.array( dat_list )
+	# 	except:
+	# 		print( 'processing cru re-gridding in serial due to multiprocessing issues...' )
+	# 		dat = np.array([ wrap( **i ) for i in args ])
 
-		lons = self._lonpc
-		if self._rotated == True: # rotate it back
-			dat, lons = self.rotate( dat, lons, to_pacific=False )
+	# 	lons = self._lonpc
+	# 	if self._rotated == True: # rotate it back
+	# 		dat, lons = self.rotate( dat, lons, to_pacific=False )
 				
-		# place back into a new xarray.Dataset object for further processing
-		self.ds = self.ds.update( { self.variable:( ['time','lat','lon'], dat ) } )
-		print( 'ds interpolated updated into self.ds' )
-		return 1
+	# 	# place back into a new xarray.Dataset object for further processing
+	# 	self.ds = self.ds.update( { self.variable:( ['time','lat','lon'], dat ) } )
+	# 	print( 'ds interpolated updated into self.ds' )
+	# 	return 1
